% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipes.R
\name{iam_create_user_and_add_to_group}
\alias{iam_create_user_and_add_to_group}
\title{create new user(s) and immediately add them to group(s)
every user is added to every group}
\usage{
iam_create_user_and_add_to_group(iam = paws::iam(), username, groupname)
}
\arguments{
\item{iam}{the iam object from paws}

\item{username}{vector of usernames}

\item{groupname}{vector of groupnames}
}
\value{
list with creation results and group_add results
}
\description{
create new user(s) and immediately add them to group(s)
every user is added to every group
}
\examples{
if(FALSE){ paws::ec2()$describe_instances() |> tidypaws() }
}
