% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipes.R
\name{s3_get_key_url}
\alias{s3_get_key_url}
\title{get the url of a key in an s3 bucket}
\usage{
s3_get_key_url(bucket, key)
}
\arguments{
\item{bucket}{the s3 bucket to query}

\item{key}{the key you want}
}
\description{
get the url of a key in an s3 bucket
}
