% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidypaws}
\alias{tidypaws}
\title{transform paws results to a tidyr format.}
\usage{
tidypaws(pawsresult, compact = T)
}
\arguments{
\item{pawsresult}{result of all paws function}

\item{compact}{whether to remove empty tables}
}
\value{
a list of tibbles
}
\description{
transform paws results to a tidyr format.
}
\examples{
if(FALSE){ paws::ec2()$describe_instances() |> tidypaws() }
}
